+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Git Repo for the project at [@Luzkan](https://github.com/Luzkan) in [@ClinicDB](
   - Editing unconfirmed visit by patient
 - (Bug) Start/End hour of a doc doesn't display for a patient looking up for it
 
+## [1.3.0] - 2019-01-13
+### Added
+- Button for patients to show up all names of the doctors in database
+- Table to view actual upcoming visits for logged in patient
+
 ## [1.2.2] - 2019-01-13
 ### Added
 - Now Password will be covered up